+++ packages/driver/src/cypress/error_messages.ts
@@ -2412,7 +2412,10 @@ export default {
         - you don't have internet access
         - you forgot to run / boot your web server
         - your web server isn't accessible
-        - you have weird network configuration settings on your computer`,
+        - you have weird network configuration settings on your computer
+        - your web server took too long to send its first response
+
+      If the error above is 'ESOCKETTIMEDOUT', your web server did not send a response within the 'responseTimeout' (30000ms by default). Note that 'pageLoadTimeout' (and the 'timeout' option on 'cy.visit()') does not apply to this initial response - increase 'responseTimeout' in your config or pass it to 'cy.visit()' instead: cy.visit(url, { responseTimeout: 60000 }).`,
     loading_file_failed (obj) {
       return stripIndent`
         ${cmd('visit')} failed trying to load: